requirements.txt
opencv-python
deepface
mediapipe
pygame
pywebview
requests
numpy